requirements.txt
matplotlib==3.4.3
GDAL==3.3.2
Shapely==1.7.1
geopandas==0.10.2
albumentations==1.1.0
segmentation_models_pytorch==0.2.0
torchvision==0.10.1+cu102
pandas==1.3.3
tqdm==4.62.3
natsort==7.1.1
rasterio==1.2.10
torch==1.9.1+cu102
numpy==1.21.2
torchsummary==1.5.1
osgeo==0.0.1
Pillow==8.4.0
scikit_learn==1.0
